Added ProgressBarGUI class, and modified saveAs().
ProgressBar now shows properly for all files within the main sketch
folder, but not within sub-folders.

To Do:
1. Make ProgressBar take into account files that have already been copied
within sub-folders in the sketch, and not onlyafter the entire folder has
been copied.
2. Rename "Saving X As Y..." with "Saving <original_project> As
<new_project>"
3. The progress made for very large files is shown only after the ENTIRE
file has been copied. Modify processing.app.Base as well so that for large
files, the ProgressBar is updated in (almost) real time
4. Make "Done Saving" appear in Message Area only after the entire Save As
operation is complete